[TECH] refactor to feature-folders
