+++ src/front/app/common/components/challenge/ChallengeJumbotron.js
