+++ src/front/app/common/components/challenge/ChallengeTile.js
