+++ src/front/app/common/components/challenge/ChallengeTile.spec.js
