+++ src/front/app/common/components/challenge/ChallengeTileList.js
