+++ src/front/app/features/categories/categories.actions.spec.js
@@ -1,4 +1,4 @@
-import { getCategories, GET_CATEGORIES } from './categories';
+import { getCategories, GET_CATEGORIES } from './categories.actions';
 
 describe('action/categories', () => {
     let dispatch;
