+++ src/front/app/features/categories/components/CategoryTile.js
