+++ src/front/app/features/categories/components/CategoryTileList.js
