+++ src/front/app/features/challenges/components/ChallengeScreen.integration.spec.js
