+++ src/front/app/features/challenges/components/ChallengeScreen.spec.js
