+++ src/front/app/features/challenges/components/ChallengeScreenContainer.js
@@ -1,5 +1,5 @@
 import { connect } from 'react-redux';
-import { fetchChallenges } from './../../actions/challenges';
+import { fetchChallenges } from '../challenges.actions';
 import ChallengeScreen from './ChallengeScreen';
 
 const mapStateToProps = ({ challenges }) => ({ challenges });
