+++ src/front/app/features/challenges/components/Menu.js
