+++ src/front/app/features/challenges/components/index.scss
