+++ src/front/app/features/navigation/navigation.scss
