+++ src/front/app/routes.js
@@ -1,21 +1,21 @@
 import React from 'react';
 import { Route } from 'react-router';
-import Login from './components/Login';
-import Challenges from './components/Challenges/ChallengeScreenContainer';
-import TodoScreen from './components/TodoScreen';
-import Awards from './components/Awards';
-import Updates from './components/Updates';
-import Profile from './components/Profile';
-import Categories from './components/Categories/CategoryTileListContainer';
+import Login from './features/login';
+import CategoryTileListContainer from './features/categories/components/CategoryTileListContainer';
+import ChallengeScreenContainer from './features/challenges/components/ChallengeScreenContainer';
+import TodoScreen from './features/todo/TodoScreen';
+import Awards from './features/awards';
+import Updates from './features/updates';
+import Profile from './features/profile';
 
 export default () => (
     <Route path="/" >
         <Route path="login" component={Login} />
-        <Route path="challenges" component={Challenges} />
+        <Route path="challenges" component={ChallengeScreenContainer} />
         <Route path="awards" component={Awards} />
         <Route path="updates" component={Updates} />
         <Route path="profile" component={Profile} />
         <Route path="todo" component={TodoScreen} />
-        <Route path="categories" component={Categories} />
+        <Route path="categories" component={CategoryTileListContainer} />
     </Route>
 );
