+++ src/front/app/store/index.js
@@ -2,8 +2,8 @@ import { createStore, combineReducers } from 'redux';
 import { hashHistory } from 'react-router';
 import storeEnhancers from './enhancers';
 import { syncHistoryWithStore, routerReducer } from 'react-router-redux';
-import challenges from '../reducers/challenges';
-import categories from '../reducers/categories';
+import challenges from '../features/challenges/challenges.reducer';
+import categories from '../features/categories/categories.reducer';
 
 export default () => {
     const state = createStore(
