+++ src/front/index.scss
@@ -1,8 +1,8 @@
 @import '../../node_modules/normalize.css/normalize.css';
-@import 'app/components/todo-screen';
-@import 'app/components/Login/login-screen';
-@import './app/components/Navigation/navigation';
-@import './app/components/Challenges/index.scss';
+@import 'app/features/todo/todo-screen';
+@import 'app/features/login/login-screen';
+@import './app/features/navigation/navigation';
+@import 'app/features/challenges/components/index';
 
 * {
   box-sizing: border-box;
